Moved old voc management to legacy
