+++ legacy/load_vocabulary.py
