+++ user_data/Denis_vocabulary_mastery.json
@@ -900,8 +900,8 @@
         "total_attempts": 6
     },
     "226": {
-        "correct_attempts": 9,
-        "total_attempts": 9
+        "correct_attempts": 10,
+        "total_attempts": 10
     },
     "227": {
         "correct_attempts": 13,
@@ -1001,7 +1001,7 @@
     },
     "251": {
         "correct_attempts": 3,
-        "total_attempts": 3
+        "total_attempts": 4
     },
     "252": {
         "correct_attempts": 4,
@@ -1076,8 +1076,8 @@
         "total_attempts": 4
     },
     "270": {
-        "correct_attempts": 1,
-        "total_attempts": 1
+        "correct_attempts": 2,
+        "total_attempts": 2
     },
     "271": {
         "correct_attempts": 2,
@@ -1112,8 +1112,8 @@
         "total_attempts": 1
     },
     "279": {
-        "correct_attempts": 0,
-        "total_attempts": 0
+        "correct_attempts": 1,
+        "total_attempts": 1
     },
     "280": {
         "correct_attempts": 2,
@@ -1128,8 +1128,8 @@
         "total_attempts": 1
     },
     "283": {
-        "correct_attempts": 0,
-        "total_attempts": 0
+        "correct_attempts": 1,
+        "total_attempts": 1
     },
     "284": {
         "correct_attempts": 1,
@@ -1140,8 +1140,8 @@
         "total_attempts": 1
     },
     "286": {
-        "correct_attempts": 0,
-        "total_attempts": 0
+        "correct_attempts": 1,
+        "total_attempts": 1
     },
     "287": {
         "correct_attempts": 1,
@@ -1160,16 +1160,16 @@
         "total_attempts": 1
     },
     "291": {
-        "correct_attempts": 0,
-        "total_attempts": 0
-    },
-    "292": {
-        "correct_attempts": 0,
+        "correct_attempts": 1,
         "total_attempts": 1
     },
+    "292": {
+        "correct_attempts": 1,
+        "total_attempts": 2
+    },
     "293": {
-        "correct_attempts": 0,
-        "total_attempts": 0
+        "correct_attempts": 1,
+        "total_attempts": 1
     },
     "294": {
         "correct_attempts": 0,
@@ -1213,7 +1213,7 @@
     },
     "304": {
         "correct_attempts": 0,
-        "total_attempts": 0
+        "total_attempts": 1
     },
     "305": {
         "correct_attempts": 1,
