+++ user_data/player_scores.json
@@ -52,8 +52,8 @@
             "hoofdstuck3": {
                 "levels": {
                     "4": {
-                        "ema": 82.30943329224829,
-                        "games_played": 18
+                        "ema": 81.73207496918621,
+                        "games_played": 19
                     },
                     "1": {
                         "ema": 92.5,
@@ -98,7 +98,7 @@
                 }
             }
         },
-        "last_played": "2024-12-02 13:28:21"
+        "last_played": "2024-12-02 13:32:56"
     },
     "Z\u00e9lie": {
         "vocabularies": {
